Implement "parallelism"
Argon2 can be configured in a way that allows parts of the state to be computed
in parallel. We don't actually *do* them in parallel yet, but at least we get
the right result now.

This is the branchiest crypto code i have ever written. And possibly the
ugliest. Will clean up in a future pass.